+++ CHANGELOG.md
@@ -5,8 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [v0.1.1] - 2024-01-17
+
+### Added
+
+- New base rules to ensure that:
+  - values are set in variables instead of in properties;
+  - no important is used;
+  - max nesting;
+  - selector consistency.
+
 ## [v0.1.0] - 2024-01-17
 
 - First release
 
+[v0.1.1]: https://www.npmjs.com/package/@dipcode/stylelint-config/v/0.1.1
 [v0.1.0]: https://www.npmjs.com/package/@dipcode/stylelint-config/v/0.1.0